refactor: Remove framework dependencies from the application layer refactor: Refactor backup functionality, naming feat: Extract TrackableFileStream as interface, introduce DataBufferFluxTrackableFileStream; Remove Spring from DownloadService feat: Refactor file replication and Game Provider interfaces feat: Add DB uniqueness constraint for FileCopy file path ci: Remove unnecessary maven-failsafe-plugin configuration from POM

+++ backend/src/main/java/dev/codesoapbox/backity/core/discovery/domain/GameContentDiscoveryProgress.java
@@ -9,7 +9,7 @@ public record GameContentDiscoveryProgress(
         GameProviderId gameProviderId,
         int percentage,
         Duration timeLeft,
-        int gamesDiscovered,
+        long gamesDiscovered,
         int gameFilesDiscovered
 ) {
